// src/types/botConfiguration.ts
export interface BotConfiguration {
  clinicId: string;
  clinicName: string;
  phoneNumber: string;
  operationHours: string;
  micrositeUrl: string;
  address: string;
  addressDescription: string;
  state: string;
  timeZone: string;
  timeZoneOffset: string;
  bookingLink: string;
  availableTimeSlots: string;
  timeIntervals: string;
  wellnessPlanPrices: string;
  calendarId: string;
  bookingCalendarId: string;
  clinicEmail: string;
  baseId: string;
  tableId: string;
  bookingWorkflow: string;
  planPrice: string;
}